Enable pvops builds for EC2
The pvops kernel comes first with SLES12 SP2 and provides
Xen HVM and Xen paravirtual operations. The paravirtual
block drivers are new to this kernel and must be addressed
in kiwi. In addition the setup of the root device in the
kernel commandline cannot be a fixed device name anymore
since the device node names are different depending which
virtualization mode is used. Therefore the root device setup
for ec2 builds is now based on the rootfs label. This
Fixes (bnc#966293)