Add part_msdos module for self build EFI image
On ARM we have a few hooks in kiwi that allow us to adapt the
resulting image to target boards. On some systems, we need to
install the firmware to hardware specified sector offsets on
the image target (sd card).

Unfortunately some systems demand firmware at sector 1 which is
where the GPT would usually reside. So we need to use an MBR partition
layout.

We can convert the GPT into an MBR using gdisk in our ARM scripts,
but at that point grub2 would have to be able to interpret the MBR
as well. So we need the "msdos_part" module included (bnc#975898)