Remove null padding on the vmware disk tag
The block of data read via dd is null padded; adding the tools data after
the padding breaks detection of the data. We need to remove the nulls (0x0)
so appends can occur adjacent to the block of strings. Also adding the
encoding statement to vmware disk tag only if not present (bnc#988087)